packages/arkham-shard-media-forensics/pyproject.toml
[project]
name = "arkham-shard-media-forensics"
version = "0.1.0"
description = "Media metadata extraction and forensic analysis for images"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "arkham-frame>=0.1.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "exifread>=3.0.0",
    "pysolar>=0.10",
    "scipy>=1.10.0",
    "numpy>=1.20.0",
    "structlog>=21.0.0",
    "c2pa-python>=0.4.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "black", "mypy"]

[project.entry-points."arkham.shards"]
media-forensics = "arkham_shard_media_forensics:MediaForensicsShard"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
